# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warp Storage Engine API
  description: |
    S3-compatible storage proxy with additional features for authentication, sharing, and monitoring.
    
    This API is mostly compatible with AWS S3 REST API. For S3 operations, refer to the 
    [AWS S3 API documentation](https://docs.aws.amazon.com/AmazonS3/latest/API/Welcome.html).
    
    This documentation covers the custom endpoints specific to Warp Storage Engine.
  version: 1.0.0
  contact:
    name: Warp Storage Engine
    url: https://github.com/prysmsh/warp-storage-engine
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
security:
  - BasicAuth: []
  - AWSSignatureV2: []
  - AWSSignatureV4: []
  - DatabaseAuth: []
...
paths:
...
  /{bucket}:
    get:
      tags:
        - S3 Operations
      summary: List objects in bucket
      description: |
        Lists objects in a bucket. Compatible with S3 ListObjectsV2 API.
        
        For detailed documentation, see [AWS S3 ListObjectsV2](https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectsV2.html)
      operationId: listObjects
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
        - name: list-type
          in: query
          schema:
            type: string
            enum: ["2"]
          description: Use "2" for ListObjectsV2
        - name: prefix
          in: query
          schema:
            type: string
          description: Limits response to keys that begin with the specified prefix
        - name: delimiter
          in: query
          schema:
            type: string
          description: Character used to group keys
        - name: max-keys
          in: query
          schema:
            type: integer
            default: 1000
          description: Maximum number of keys returned
        - name: continuation-token
          in: query
          schema:
            type: string
          description: Token for pagination
      responses:
        '200':
          description: List of objects
          content:
            application/xml:
              schema:
                type: string
                example: |
                  <?xml version="1.0" encoding="UTF-8"?>
                  <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
                    <Name>bucket-name</Name>
                    <Prefix/>
                    <MaxKeys>1000</MaxKeys>
                    <IsTruncated>false</IsTruncated>
                    <Contents>
                      <Key>file.txt</Key>
                      <LastModified>2023-01-01T00:00:00.000Z</LastModified>
                      <Size>1024</Size>
                      <StorageClass>STANDARD</StorageClass>
                    </Contents>
                  </ListBucketResult>
        '403':
          description: Access denied
        '404':
          description: Bucket not found

    put:
      tags:
        - S3 Operations
      summary: Create bucket
      description: Creates a new S3 bucket
      operationId: createBucket
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
      responses:
        '200':
          description: Bucket created successfully
        '409':
          description: Bucket already exists

    delete:
      tags:
        - S3 Operations
      summary: Delete bucket
      description: Deletes an empty S3 bucket
      operationId: deleteBucket
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
      responses:
        '204':
          description: Bucket deleted successfully
        '409':
          description: Bucket not empty
  /{bucket}/{key}:
    get:
      tags:
        - S3 Operations
      summary: Get object
      description: Downloads an object from S3
      operationId: getObject
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Object key
      responses:
        '200':
          description: Object content
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found

    put:
      tags:
        - S3 Operations
      summary: Put object
      description: Uploads an object to S3
      operationId: putObject
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Object key
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Object uploaded successfully
        '413':
          description: Request entity too large

    delete:
      tags:
        - S3 Operations
      summary: Delete object
      description: Deletes an object from S3
      operationId: deleteObject
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Object key
      responses:
        '204':
          description: Object deleted successfully

    head:
      tags:
        - S3 Operations
      summary: Head object
      description: Retrieves metadata about an object without downloading it
      operationId: headObject
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
          description: Bucket name
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Object key
      responses:
        '200':
          description: Object metadata
          headers:
            Content-Length:
              schema:
                type: integer
            Content-Type:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
        '404':
          description: Object not found
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication using username and password

    AWSSignatureV2:
      type: apiKey
      in: header
      name: Authorization
      description: |
        AWS Signature Version 2 authentication.
        Format: `AWS AccessKeyId:Signature`

    AWSSignatureV4:
      type: apiKey
      in: header
      name: Authorization
      description: |
        AWS Signature Version 4 authentication.
        Format: `AWS4-HMAC-SHA256 Credential=AccessKeyId/Date/Region/Service/aws4_request, SignedHeaders=Headers, Signature=Signature`

    DatabaseAuth:
      type: http
      scheme: basic
      description: |
        Database-backed authentication. Can be used with Basic Auth or AWS Signatures.
        When using Basic Auth, the username is the access key and password is checked against the database.
        When using AWS Signatures, credentials are validated against the database.